requirements.txt
sentence-transformers
scikit-learn
pyyaml
